Use keccak hash for payment IDs instead of uint256
Previously, we used a uint256, which was directly emitted by the InvoiceNFT, to de-dupe receivedPayment calls. Using a bytes32 is more flexible, since it supports multiple variable types of payment IDs, and also supports multiple NFTs, which the previous solution didn't. (It's impossible not to have uint256 collisions without having each NFT talk to each other or share a counter, which just makes no sense).